backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pymysql==1.1.1
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.8
bcrypt==3.2.2
pypdf==4.3.1